extract haproxy frontend+backend config to separate resource
